+++ milestone-2-pipex/error.c
@@ -6,7 +6,7 @@
 /*   By: elagouch <elagouch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/27 18:17:07 by elagouch          #+#    #+#             */
-/*   Updated: 2025/01/27 18:17:28 by elagouch         ###   ########.fr       */
+/*   Updated: 2025/01/28 15:57:52 by elagouch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ void	perror_and_exit(void)
 /**
  * @brief	Set errno, print the error message and exit
  */
-void	perror_errno_and_exit(int err)
+void	perror_errno_and_exit(size_t err)
 {
 	errno = err;
 	perror_and_exit();
